fix: pool mask overflow (uint8/16/32/64 tiers + shift fix)

+++ test/cases/phase11/pool/alloc-c-output/expected.c
@@ -8,8 +8,8 @@ static uint8_t _token_pool_mask = 0;
 
 static opt_ref_Token Token_alloc(void) {
     for (int _i = 0; _i < 2; _i++) {
-        if (!(_token_pool_mask & (1 << _i))) {
-            _token_pool_mask |= (1 << _i);
+        if (!(_token_pool_mask & ((uint8_t)1 << _i))) {
+            _token_pool_mask |= ((uint8_t)1 << _i);
             return (opt_ref_Token){true, &_token_pool[_i], _i};
         }
     }
